Remove unnecessary modules
