--- app/src/marketplace/modules/createContractProduct/constants.js
@@ -1,6 +1,0 @@
-// @flow
-
-export const CREATE_CONTRACT_PRODUCT_REQUEST: string = 'marketplace/createContractProduct/CREATE_CONTRACT_PRODUCT_REQUEST'
-export const CREATE_CONTRACT_PRODUCT_SUCCESS: string = 'marketplace/createContractProduct/CREATE_CONTRACT_PRODUCT_SUCCESS'
-export const CREATE_CONTRACT_PRODUCT_FAILURE: string = 'marketplace/createContractProduct/CREATE_CONTRACT_PRODUCT_FAILURE'
-export const RECEIVE_CREATE_CONTRACT_PRODUCT_HASH: string = 'marketplace/createContractProduct/RECEIVE_CREATE_CONTRACT_PRODUCT_HASH'
