--- app/src/marketplace/modules/publish/constants.js
@@ -1,14 +1,0 @@
-// @flow
-
-export const DEPLOY_PRODUCT_REQUEST: string = 'marketplace/publish/DEPLOY_PRODUCT_REQUEST'
-export const DEPLOY_PRODUCT_SUCCESS: string = 'marketplace/publish/DEPLOY_PRODUCT_SUCCESS'
-export const DEPLOY_PRODUCT_FAILURE: string = 'marketplace/publish/DEPLOY_PRODUCT_FAILURE'
-export const RECEIVE_DEPLOY_PRODUCT_HASH: string = 'marketplace/publish/RECEIVE_DEPLOY_PRODUCT_HASH'
-
-export const POST_DEPLOY_FREE_PRODUCT_REQUEST: string = 'marketplace/publish/POST_DEPLOY_FREE_PRODUCT_REQUEST'
-export const POST_DEPLOY_FREE_PRODUCT_SUCCESS: string = 'marketplace/publish/POST_DEPLOY_FREE_PRODUCT_SUCCESS'
-export const POST_DEPLOY_FREE_PRODUCT_FAILURE: string = 'marketplace/publish/POST_DEPLOY_FREE_PRODUCT_FAILURE'
-
-export const SET_PRODUCT_DEPLOYING_REQUEST: string = 'marketplace/publish/SET_PRODUCT_DEPLOYING_REQUEST'
-export const SET_PRODUCT_DEPLOYING_SUCCESS: string = 'marketplace/publish/SET_PRODUCT_DEPLOYING_SUCCESS'
-export const SET_PRODUCT_DEPLOYING_FAILURE: string = 'marketplace/publish/SET_PRODUCT_DEPLOYING_FAILURE'
