--- app/src/marketplace/modules/publish/types.js
@@ -1,27 +1,0 @@
-// @flow
-
-import type { ErrorInUi, PayloadAction } from '$shared/flowtype/common-types'
-import type { ProductId } from '$mp/flowtype/product-types'
-import type { Hash } from '$shared/flowtype/web3-types'
-
-export type ProductIdAction = PayloadAction<{
-    id: ProductId,
-}>
-export type ProductIdActionCreator = (ProductId) => ProductIdAction
-
-export type PublishAction = PayloadAction<{
-    id: ProductId,
-}>
-export type PublishActionCreator = (id: ProductId) => PublishAction
-
-export type HashAction = PayloadAction<{
-    hash: Hash,
-}>
-export type HashActionCreator = (Hash) => HashAction
-
-export type PublishErrorAction = PayloadAction<{
-    id: ProductId,
-    error: ErrorInUi,
-}>
-export type PublishErrorActionCreator = (ProductId, ErrorInUi) => PublishErrorAction
-
