--- app/src/marketplace/modules/unpublish/constants.js
@@ -1,14 +1,0 @@
-// @flow
-
-export const UNDEPLOY_PRODUCT_REQUEST: string = 'marketplace/unpublish/UNDEPLOY_PRODUCT_REQUEST'
-export const UNDEPLOY_PRODUCT_SUCCESS: string = 'marketplace/unpublish/UNDEPLOY_PRODUCT_SUCCESS'
-export const UNDEPLOY_PRODUCT_FAILURE: string = 'marketplace/unpublish/UNDEPLOY_PRODUCT_FAILURE'
-export const RECEIVE_UNDEPLOY_PRODUCT_HASH: string = 'marketplace/unpublish/RECEIVE_UNDEPLOY_PRODUCT_HASH'
-
-export const POST_UNDEPLOY_FREE_PRODUCT_REQUEST: string = 'marketplace/unpublish/POST_UNDEPLOY_FREE_PRODUCT_REQUEST'
-export const POST_UNDEPLOY_FREE_PRODUCT_SUCCESS: string = 'marketplace/unpublish/POST_UNDEPLOY_FREE_PRODUCT_SUCCESS'
-export const POST_UNDEPLOY_FREE_PRODUCT_FAILURE: string = 'marketplace/unpublish/POST_UNDEPLOY_FREE_PRODUCT_FAILURE'
-
-export const SET_PRODUCT_UNDEPLOYING_REQUEST: string = 'marketplace/unpublish/SET_PRODUCT_UNDEPLOYING_REQUEST'
-export const SET_PRODUCT_UNDEPLOYING_SUCCESS: string = 'marketplace/unpublish/SET_PRODUCT_UNDEPLOYING_SUCCESS'
-export const SET_PRODUCT_UNDEPLOYING_FAILURE: string = 'marketplace/unpublish/SET_PRODUCT_UNDEPLOYING_FAILURE'
