--- app/src/marketplace/modules/updateContractProduct/constants.js
@@ -1,7 +1,0 @@
-// @flow
-
-export const UPDATE_CONTRACT_PRODUCT_REQUEST: string = 'marketplace/updateContractProduct/UPDATE_CONTRACT_PRODUCT_REQUEST'
-export const UPDATE_CONTRACT_PRODUCT_SUCCESS: string = 'marketplace/updateContractProduct/UPDATE_CONTRACT_PRODUCT_SUCCESS'
-export const UPDATE_CONTRACT_PRODUCT_FAILURE: string = 'marketplace/updateContractProduct/UPDATE_CONTRACT_PRODUCT_FAILURE'
-export const RECEIVE_UPDATE_CONTRACT_PRODUCT_HASH: string = 'marketplace/updateContractProduct/RECEIVE_UPDATE_CONTRACT_PRODUCT_HASH'
-export const UPDATE_CONTRACT_PRODUCT_RESET: string = 'marketplace/updateContractProduct/UPDATE_CONTRACT_PRODUCT_RESET'
