--- app/test/unit/modules/publish/constants.test.js
@@ -1,9 +1,0 @@
-import * as constants from '$mp/modules/publish/constants'
-
-describe('publish - constants', () => {
-    it('is namespaced correctly', () => {
-        Object.keys(constants).forEach((key) => {
-            expect(constants[key]).toEqual(expect.stringMatching(/^marketplace\/publish\//))
-        })
-    })
-})
