--- app/test/unit/modules/unpublish/constants.test.js
@@ -1,9 +1,0 @@
-import * as constants from '$mp/modules/unpublish/constants'
-
-describe('unpublish - constants', () => {
-    it('is namespaced correctly', () => {
-        Object.keys(constants).forEach((key) => {
-            expect(constants[key]).toEqual(expect.stringMatching(/^marketplace\/unpublish\//))
-        })
-    })
-})
