--- app/test/unit/modules/updateContractProduct/constants.test.js
@@ -1,9 +1,0 @@
-import * as constants from '$mp/modules/updateContractProduct/constants'
-
-describe('updateContractProduct - constants', () => {
-    it('is namespaced correctly', () => {
-        Object.keys(constants).forEach((key) => {
-            expect(constants[key]).toEqual(expect.stringMatching(/^marketplace\/updateContractProduct\//))
-        })
-    })
-})
